fix: cyclic dependency
