+++ addons/addon-base-raas/packages/base-raas-services/lib/environment/__tests__/environment-mount-service.test.js
@@ -76,81 +76,6 @@ describe('EnvironmentMountService', () => {
     iamService = await container.find('iamService');
   });
 
-  describe('updateStudyFileMountIPAllowList', () => {
-    it('should not call storageGatewayService if mounted studies do not have file share', async () => {
-      // BUILD
-      const studiesList = [
-        { id: 'study1', resources: [{ arn: 'study1-s3-path-arn' }] },
-        { id: 'study2', resources: [{ arn: 'study2-s3-path-arn' }] },
-      ];
-      studyService.listByIds = jest.fn().mockResolvedValueOnce(studiesList);
-      const environment = { studyIds: ['study1', 'study2'] };
-      // OPERATE
-      await service.updateStudyFileMountIPAllowList(context, environment, {});
-      // CHECK
-      expect(storageGatewayService.updateFileSharesIPAllowedList).not.toHaveBeenCalled();
-    });
-
-    it('should call storageGatewayService with correct parameter for Add IP to allow list', async () => {
-      // BUILD
-      const studiesList = [
-        { id: 'study1', resources: [{ arn: 'study1-s3-path-arn', fileShareArn: 'study1-file-share-arn' }] },
-        { id: 'study2', resources: [{ arn: 'study2-s3-path-arn', fileShareArn: 'study2-file-share-arn' }] },
-      ];
-      studyService.listByIds = jest.fn().mockResolvedValueOnce(studiesList);
-      const environment = { studyIds: ['study1', 'study2'] };
-      const ipAllowListAction = { ip: '12.23.34.45', action: 'ADD' };
-      // OPERATE
-      await service.updateStudyFileMountIPAllowList(context, environment, ipAllowListAction);
-      // CHECK
-      expect(storageGatewayService.updateFileSharesIPAllowedList).toHaveBeenCalledWith(
-        ['study1-file-share-arn', 'study2-file-share-arn'],
-        '12.23.34.45',
-        'ADD',
-      );
-    });
-
-    it('should not call storageGatewayService when ip it not in environment or ipAllowListAction', async () => {
-      // BUILD
-      const studiesList = [
-        { id: 'study1', resources: [{ arn: 'study1-s3-path-arn', fileShareArn: 'study1-file-share-arn' }] },
-        { id: 'study2', resources: [{ arn: 'study2-s3-path-arn', fileShareArn: 'study2-file-share-arn' }] },
-      ];
-      studyService.listByIds = jest.fn().mockResolvedValueOnce(studiesList);
-      const environment = {
-        studyIds: ['study1', 'study2'],
-        outputs: [{ OutputKey: 'Ec2WorkspaceInstanceId', OutputValue: 'some-ec2-instance-id' }],
-      };
-      const ipAllowListAction = { action: 'REMOVE' };
-      // OPERATE
-      await service.updateStudyFileMountIPAllowList(context, environment, ipAllowListAction);
-      // CHECK
-      expect(storageGatewayService.updateFileSharesIPAllowedList).not.toHaveBeenCalled();
-    });
-
-    it('should call storageGatewayService with correct parameter for REMOVE IP to allow list', async () => {
-      // BUILD
-      const studiesList = [
-        { id: 'study1', resources: [{ arn: 'study1-s3-path-arn', fileShareArn: 'study1-file-share-arn' }] },
-        { id: 'study2', resources: [{ arn: 'study2-s3-path-arn', fileShareArn: 'study2-file-share-arn' }] },
-      ];
-      studyService.listByIds = jest.fn().mockResolvedValueOnce(studiesList);
-      const environment = {
-        studyIds: ['study1', 'study2'],
-        outputs: [{ OutputKey: 'Ec2WorkspacePublicIp', OutputValue: '34.45.56.67' }],
-      };
-      const ipAllowListAction = { action: 'REMOVE' };
-      // OPERATE
-      await service.updateStudyFileMountIPAllowList(context, environment, ipAllowListAction);
-      // CHECK
-      expect(storageGatewayService.updateFileSharesIPAllowedList).toHaveBeenCalledWith(
-        ['study1-file-share-arn', 'study2-file-share-arn'],
-        '34.45.56.67',
-        'REMOVE',
-      );
-    });
-  });
-
   describe('Update paths', () => {
     it('should call nothing if all are admin changes', async () => {
       // BUILD
