+++ addons/addon-base-raas/packages/base-raas-services/lib/environment/environment-mount-service.js
@@ -102,43 +102,6 @@ class EnvironmentMountService extends Service {
     return this._updateResourcePolicies({ updateAwsPrincipals, workspaceRoleArn, s3Prefixes });
   }
 
-  /**
-   * This method is triggered when EC2 Linux, Windows and RStudio Start / Stop
-   * And when any workspace is terminated
-   * @param requestContext
-   * @param existingEnvironment: environment that's being stopped / started or terminated
-   * @param ipAllowListAction: One required field action and one optional field ip
-   * @return {Promise<void>}
-   */
-  async updateStudyFileMountIPAllowList(requestContext, existingEnvironment, ipAllowListAction) {
-    const [storageGatewayService, studyService] = await this.service(['storageGatewayService', 'studyService']);
-    // Check if the mounted study is using StorageGateway
-    const studiesList = await studyService.listByIds(
-      requestContext,
-      existingEnvironment.studyIds.map(id => {
-        return { id };
-      }),
-    );
-
-    // If yes, get the file share ARNs and call to update IP allow list
-    const fileShareARNs = studiesList.map(study => study.resources[0].fileShareArn).filter(arn => !_.isUndefined(arn));
-    if (!_.isEmpty(fileShareARNs)) {
-      let ip;
-      // If IP is in ipAllowListAction, use that, if not, find it in existingEnvironment
-      if ('ip' in ipAllowListAction) {
-        ip = ipAllowListAction.ip;
-      } else {
-        ip = existingEnvironment.outputs.filter(output => output.OutputKey === 'Ec2WorkspacePublicIp');
-        // When terminating a product that's not EC2 based, do nothing
-        if (_.isEmpty(ip)) {
-          return;
-        }
-        ip = ip[0].OutputValue;
-      }
-      await storageGatewayService.updateFileSharesIPAllowedList(fileShareARNs, ip, ipAllowListAction.action);
-    }
-  }
-
   async _updateResourcePolicies({ updateAwsPrincipals, workspaceRoleArn, s3Prefixes }) {
     if (s3Prefixes.length === 0) {
       return;
